added usage counter logic and waste collector to block model/reg; refactod block lifecicle & logic

+++ bstore/bssrv/bsblock/bsrec.go
@@ -12,13 +12,17 @@ import (
     "os"
     "errors"
     "encoding/hex"
-    "crypto/md5"
+    "crypto/sha1"
+    "math/rand"
     "time"
+
+    "github.com/minio/highwayhash"
 
     "ndstore/bstore/bssrv/bsbreg"
     "ndstore/dsrpc"
     "ndstore/dscom"
     "ndstore/dserr"
+    "ndstore/dslog"
 )
 
 const blockFileExt string = ".blk"
@@ -28,6 +32,7 @@ type Store struct {
     reg *bsbreg.Reg
     dirPerm   fs.FileMode
     filePerm  fs.FileMode
+    wasteChan chan byte
 }
 
 func NewStore(dataRoot string, reg *bsbreg.Reg) *Store {
@@ -36,6 +41,7 @@ func NewStore(dataRoot string, reg *bsbreg.Reg) *Store {
     store.reg       = reg
     store.dirPerm   = 0755
     store.filePerm  = 0644
+    store.wasteChan = make(chan byte, 1024)
     return &store
 }
 
@@ -47,18 +53,29 @@ func (store *Store) SetFilePerm(filePerm fs.FileMode) {
     store.filePerm = filePerm
 }
 
-func (store *Store) SaveBlock(fileId, batchId, blockId, blockSize, dataSize int64, blockReader io.Reader,
-                                    binSize int64, blockType, hashAlg, hashInit, hashSum string) error {
-    var err error
 
-    //blockExists, err := store.reg.BlockDescrExists(fileId, batchId, blockId, blockType)
-    //if err != nil {
-        //return dserr.Err(err)
-    //}
-    //if blockExists {
-        //err = errors.New("block yet exists")
-        //return dserr.Err(err)
-    //}
+func (store *Store) SaveBlock(fileId, batchId, blockId int64,  blockType string, blockSize, dataSize int64,
+                            hashAlg, hashInit, hashSum string, blockReader io.Reader, binSize int64) error {
+
+    var err error
+    const uCounter int64 = 1
+
+    exists, used, _, _, err := store.reg.GetBlockFilePath(fileId, batchId, blockId, blockType)
+
+    if err != nil {
+        return dserr.Err(err)
+    }
+    if exists && used {
+        err = errors.New("block yet exists")
+        return dserr.Err(err)
+    }
+
+    if exists && !used {
+        err = store.dropBlock(fileId, batchId, blockId, blockType)
+        if err != nil {
+            return dserr.Err(err)
+        }
+    }
 
     err = validateFileId(fileId)
     if err != nil {
@@ -75,7 +92,8 @@ func (store *Store) SaveBlock(fileId, batchId, blockId, blockSize, dataSize int6
         return dserr.Err(err)
     }
 
-    fileName := MakeBlockName(fileId, batchId, blockId)
+
+    fileName := MakeBlockName(fileId, batchId, blockId, blockType)
     subdirName := MakeDirName(fileName)
     dirPath := filepath.Join(store.dataRoot, subdirName)
     os.MkdirAll(dirPath, store.dirPerm)
@@ -87,13 +105,32 @@ func (store *Store) SaveBlock(fileId, batchId, blockId, blockSize, dataSize int6
     if err != nil {
         return dserr.Err(err)
     }
-    _, err = dsrpc.CopyBytes(blockReader, blockFile, binSize)
-    if err != nil {
-        return dserr.Err(err)
+
+    switch {
+        case len(hashSum) == 0:
+            hashIBytes := make([]byte, 32)
+            rand.Read(hashIBytes)
+            hasher, _ := highwayhash.New(hashIBytes)
+
+            multiWriter := io.MultiWriter(blockFile, hasher)
+            _, err = dsrpc.CopyBytes(blockReader, multiWriter, binSize)
+            if err != nil {
+                return dserr.Err(err)
+            }
+            hashBytes := hasher.Sum(nil)
+            hashSum = hex.EncodeToString(hashBytes)
+            hashInit = hex.EncodeToString(hashIBytes)
+            hashAlg = dscom.HashTypeHW
+        default:
+           _, err = dsrpc.CopyBytes(blockReader, blockFile, binSize)
+            if err != nil {
+                return dserr.Err(err)
+            }
     }
+
     filePath := filepath.Join(subdirName, fileName)
-    err = store.reg.AddBlockDescr(fileId, batchId, blockId, blockSize, dataSize, filePath,
-                                                            blockType, hashAlg, hashInit, hashSum)
+    err = store.reg.AddBlockDescr(fileId, batchId, blockId, uCounter, blockSize, dataSize,
+                                                    filePath, blockType, hashAlg, hashInit, hashSum)
 
     if err != nil {
         os.Remove(filePath)
@@ -106,10 +143,16 @@ func (store *Store) BlockExists(fileId, batchId, blockId int64, blockType string
     var err error
     var filePath string
     var exists bool
-    filePath, blockSize, err := store.reg.GetBlockFilePath(fileId, batchId, blockId, blockType)
+    exists, used, filePath, blockSize, err := store.reg.GetBlockFilePath(fileId, batchId, blockId, blockType)
     if err != nil {
         return exists, blockSize, dserr.Err(err)
     }
+    if !exists || (exists && !used) {
+        exists = false
+        return exists, blockSize, dserr.Err(err)
+    }
+
+    exists = false
     filePath = filepath.Join(store.dataRoot, filePath)
     blockFile, err := os.OpenFile(filePath, os.O_RDONLY, 0)
     defer blockFile.Close()
@@ -124,10 +167,15 @@ func (store *Store) CheckBlock(fileId, batchId, blockId int64, blockType string)
     var err error
     var filePath string
     var correct bool
-    filePath, dataSize, err := store.reg.GetBlockFilePath(fileId, batchId, blockId, blockType)
+    exists, used, filePath, dataSize, err := store.reg.GetBlockFilePath(fileId, batchId, blockId, blockType)
     if err != nil {
         return correct, dserr.Err(err)
     }
+
+    if !exists || (exists && !used) {
+        return correct, dserr.Err(err)
+    }
+
     filePath = filepath.Join(store.dataRoot, filePath)
     blockFile, err := os.OpenFile(filePath, os.O_RDONLY, 0)
     defer blockFile.Close()
@@ -149,14 +197,18 @@ func (store *Store) CheckBlock(fileId, batchId, blockId int64, blockType string)
     return correct, dserr.Err(err)
 }
 
-func (store *Store) LoadBlock(fileId, batchId, blockId int64,
-                                                    blockWriter io.Writer, blockType string) error {
+func (store *Store) LoadBlock(fileId, batchId, blockId int64, blockType string, blockWriter io.Writer) error {
     var err error
     var filePath string
-    filePath, blockSize, err := store.reg.GetBlockFilePath(fileId, batchId, blockId, blockType)
+    exists, used, filePath, blockSize, err := store.reg.GetBlockFilePath(fileId, batchId, blockId, blockType)
     if err != nil {
         return dserr.Err(err)
     }
+    if !exists || !used {
+        err = errors.New("block not exists")
+        return dserr.Err(err)
+    }
+
     filePath = filepath.Join(store.dataRoot, filePath)
     blockFile, err := os.OpenFile(filePath, os.O_RDONLY, 0)
     defer blockFile.Close()
@@ -170,22 +222,91 @@ func (store *Store) LoadBlock(fileId, batchId, blockId int64,
     return dserr.Err(err)
 }
 
-func (store *Store) DeleteBlock(fileId, batchId, blockId int64, blockType string) error {
+
+func (store *Store) WasteCollector() {
+    for {
+        exists, bl, err := store.reg.GetUnusedDescr()
+        if exists && err == nil {
+            err = store.dropBlock(bl.FileId, bl.BatchId, bl.BlockId, bl.BlockType)
+            if err != nil {
+                dslog.LogDebug("delete waste block err:", dserr.Err(err))
+            }
+            continue
+        }
+        select {
+            case <- store.wasteChan:
+            case <-time.After(time.Second * 10):
+        }
+    }
+}
+
+func (store *Store) pushWC() {
+    if cap(store.wasteChan) - len(store.wasteChan) > 1 {
+        store.wasteChan <- 0xff
+    }
+}
+
+func (store *Store) dropBlock(fileId, batchId, blockId  int64, blockType string) error {
     var err error
     var filePath string
-    filePath, _, err = store.reg.GetBlockFilePath(fileId, batchId, blockId, blockType)
+    exists, used, filePath, _, _ := store.reg.GetBlockFilePath(fileId, batchId, blockId, blockType)
     if err != nil {
         return dserr.Err(err)
     }
-    filePath = filepath.Join(store.dataRoot, filePath)
-    err = os.Remove(filePath)
+    if exists && used {
+        return dserr.Err(err)
+    }
+    if !exists  {
+        return dserr.Err(err)
+    }
+
+    if len(filePath) > 0 {
+        filePath = filepath.Join(store.dataRoot, filePath)
+        err = os.Remove(filePath)
+        if err != nil {
+            return dserr.Err(err)
+        }
+    }
+    err = store.reg.DropBlockDescr(fileId, batchId, blockId, blockType)
     if err != nil {
         return dserr.Err(err)
     }
-    err = store.reg.DeleteBlockDescr(fileId, batchId, blockId, blockType)
+    return dserr.Err(err)
+}
+
+func (store *Store) DeleteBlock(fileId, batchId, blockId int64, blockType string) error {
+    var err error
+    err = store.reg.DecDescrUCounter(fileId, batchId, blockId, blockType)
+    defer store.pushWC()
     if err != nil {
         return dserr.Err(err)
     }
+    return dserr.Err(err)
+}
+
+
+func (store *Store) PurgeAll() error {
+    var err error
+    var filePath string
+    blocks, err := store.reg.ListBlockDescrs()
+    if err != nil {
+        return dserr.Err(err)
+    }
+    for _, block := range blocks {
+        filePath = filepath.Join(store.dataRoot, block.FilePath)
+        _ = os.Remove(filePath)
+        //if err != nil {
+        //    return dserr.Err(err)
+        //}
+        err = store.reg.DropBlockDescr(block.FileId, block.BatchId, block.BlockId, block.BlockType)
+        if err != nil {
+            return dserr.Err(err)
+        }
+    }
+    //err = store.reg.PurgeAll()
+    //if err != nil {
+    //    return dserr.Err(err)
+    //}
     return dserr.Err(err)
 }
 
@@ -198,10 +319,9 @@ func (store *Store) ListBlocks() ([]*dscom.BlockDescr, error) {
     return blocks, dserr.Err(err)
 }
 
-func MakeBlockName(fileId, batchId, blockId int64) string {
-    ts := time.Now().UnixNano()
-    origin := fmt.Sprintf("%020d-%020d-%020d-%020d", fileId, batchId, blockId, ts)
-    hasher := md5.New()
+func MakeBlockName(fileId, batchId, blockId int64, blockType string) string {
+    origin := fmt.Sprintf("%020d-%020d-%020d-%020d-%s", fileId, batchId, blockId, blockType)
+    hasher := sha1.New()
     hasher.Write([]byte(origin))
     hashSum := hasher.Sum(nil)
     hashHex := hex.EncodeToString(hashSum)
@@ -210,7 +330,7 @@ func MakeBlockName(fileId, batchId, blockId int64) string {
 }
 
 func MakeDirName(fileName string) string {
-    hasher := md5.New()
+    hasher := sha1.New()
     hasher.Write([]byte(fileName))
     hashSum := hasher.Sum(nil)
     hashHex := make([]byte, hex.EncodedLen(len(hashSum)))
